brite_payments/requirements.txt
django==4.2.9
djangorestframework==3.14.0
requests==2.31.0